requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
cloudinary==1.42.1
defusedxml==0.7.1
dj-rest-auth==2.1.9
Django==5.1.5
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
pillow==11.1.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
